feat: add MyEscrows component with tabbed view and loading state
